+++ web/wp-content/plugins/image-box/package.json
@@ -21,13 +21,19 @@
 		"test:unit": "wp-scripts test-unit-js"
 	},
 	"devDependencies": {
-		"@wordpress/block-editor": "^6.2.0",
-		"@wordpress/blocks": "^10.0.0",
-		"@wordpress/components": "^14.2.0",
-		"@wordpress/data": "^5.2.0",
-		"@wordpress/element": "^3.2.0",
-		"@wordpress/i18n": "^4.2.0",
-		"@wordpress/rich-text": "^4.2.0",
-		"@wordpress/scripts": "^17.0.0"
-	}
+		"@wordpress/block-editor": "^8.1.0",
+		"@wordpress/blocks": "^11.2.0",
+		"@wordpress/components": "^19.3.0",
+		"@wordpress/data": "^6.2.0",
+		"@wordpress/element": "^4.1.0",
+		"@wordpress/i18n": "^4.3.0",
+		"@wordpress/rich-text": "^5.1.0",
+		"@wordpress/scripts": "^20.0.2",
+		"autoprefixer": "^10.4.2",
+		"classnames": "^2.3.1"
+	},
+	"browserslist": [
+		"defaults",
+		"not IE 11"
+	]
 }
