+++ web/wp-content/plugins/image-box/src/editor.scss
@@ -109,11 +109,10 @@
 		align-items: center;
 		min-height: 20px;
 
-		.block-editor-block-list__block .block-editor-url-input input[type="text"],
-		.block-editor-url-input input[type="text"],
-		.components-popover .block-editor-url-input input[type="text"] {
+		.block-editor-block-list__block .block-editor-url-input input[type='text'],
+		.block-editor-url-input input[type='text'],
+		.components-popover .block-editor-url-input input[type='text'] {
 			width: 100%;
-			border: 1px;
 			padding: 6px 8px;
 			box-shadow: 0 0 0 transparent;
 			transition: box-shadow 0.1s linear;
